requirements.txt
# GIA_INFOSYS 문서 파싱 환경 구축
# Phase 2-1, 과업 1: 문서 파싱 환경 구축

# 문서 파싱 라이브러리
python-docx==0.8.11
python-pptx==0.6.21
PyMuPDF==1.23.8

# Notion API 연동
notion-client==2.2.1

# 환경 변수 관리
python-dotenv==1.0.0

# HTTP 요청
requests==2.31.0

# 데이터 처리
pandas==2.1.4

# 기타 유틸리티
python-dateutil==2.8.2
